refactor(solver): split constraint predicate helpers
